Added the ability to pass an additional inclusion predicate to the AddSerilog configuration method.
