+++ Directory.Build.props
@@ -4,7 +4,7 @@
 		<TargetFrameworks>net8.0;net9.0;net10.0</TargetFrameworks>
 		
 		<!-- NuGet -->
-		<Version>5.4.3</Version>
+		<Version>5.4.4</Version>
 		<AssemblyVersion>5.4.0</AssemblyVersion>
 		<FileVersion>5.4.0</FileVersion>
 		<Authors>Jon Sagara</Authors>
